day3_phishing_scraper/requirements.txt
requests
beautifulsoup4
googlesearch-python
tldextract
pandas
